requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
beautifulsoup4==4.12.2
azure-ai-inference==0.1.0b1
azure-core==1.29.5
pdfplumber==0.9.0
python-docx==0.8.11
Pillow==10.0.1
pytesseract==0.3.10
Flask-WTF==1.1.1
WTForms==3.0.1
